Refactor Haskell flake
The advent of code day flake was awkward because it was way too many
files and machinery for something so simple. The only thing you need is
some GHC packages in the root and then each day is maybe one or two
haskell files loaded with ghc --package-db for part 1 and part 2.